Updated the bulk of the code that does all provisioning from a client. There is some parsed functionality to ease sorting through the changes and updates needed in the future.
